CRUD_admin
сделаны ручки для CRUD через бота, реализована админ панель

+++ menu/middlewares.py
@@ -1,8 +1,10 @@
 import asyncio
 from aiogram import BaseMiddleware
-from aiogram.types import Message, TelegramObject, CallbackQuery
+from aiogram.types import Message, TelegramObject
 from aiogram.exceptions import TelegramRetryAfter
 from typing import Callable, Dict, Any, Awaitable
+
+from menu.database.requests import check_status
 
 class AntiSpamMiddleware(BaseMiddleware):
     """
@@ -45,26 +47,28 @@ class AntiSpamMiddleware(BaseMiddleware):
         """
         
         user_id = message.from_user.id
+        flag = await check_status(user_id)
+        if not flag:
+            # Проверка, отправлял ли пользователь сообщения ранее
+            if user_id in self.cache:
+                last_message_time = self.cache[user_id]
+                time_passed = message.date.timestamp() - last_message_time
 
-        # Проверка, отправлял ли пользователь сообщения ранее
-        if user_id in self.cache:
-            last_message_time = self.cache[user_id]
-            time_passed = message.date.timestamp() - last_message_time
-
-            # Проверка на превышение лимита сообщений
-            if time_passed < self.cooldown:
-                # Отправка сообщения об ожидании
-                try:
-                    # Удаляем сообщение, если оно помечено как спам
-                    await message.delete()
-                    await message.answer(f"Пожалуйста, подождите {self.cooldown} секунд перед отправкой следующего сообщения.")
-                except TelegramRetryAfter as e:
-                    # Обработка ошибки, если сообщение отправлено слишком быстро
-                    await asyncio.sleep(e.timeout)
-                    await message.edit_text(f"Подождите {self.cooldown} секунд перед отправкой следующего сообщения.")
-                return True
-        # Обновление времени последнего сообщения пользователя
-        self.cache[user_id] = message.date.timestamp()
+                # Проверка на превышение лимита сообщений
+                if time_passed < self.cooldown:
+                    # Отправка сообщения об ожидании
+                    try:
+                        # Удаляем сообщение, если оно помечено как спам
+                        await message.delete()
+                        await message.answer(f"Пожалуйста, подождите {self.cooldown} секунд перед отправкой следующего сообщения.")
+                    except TelegramRetryAfter as e:
+                        # Обработка ошибки, если сообщение отправлено слишком быстро
+                        await asyncio.sleep(e.timeout)
+                        await message.edit_text(f"Подождите {self.cooldown} секунд перед отправкой следующего сообщения.")
+                    return True
+            # Обновление времени последнего сообщения пользователя
+            self.cache[user_id] = message.date.timestamp()
+        else: return
     
     async def __call__(self,
                        handler: Callable[[TelegramObject, Dict[str, Any]], Awaitable[Any]],
